refactor: 운동 종류 인터페이스 제거
